feat: Lotto 클래스 패키지 변경 및 내부 메서드 추가
